improve BLE client lifetime safety and code quality
- Replace raw pointers with unique_ptr in NikonBLEClient and NormalMode.

- Mark NikonBLEClient non-copyable and non-movable (engine holds hasher ref).

- Add null checks for timeChar/geoChar and clean up stale client on handshake failure.

- Log camera serial number extracted from stage 4 message.

- Rewrite Utils::hexStr with lookup table; change len type to size_t.

- Move BLEDevice::getScan() into PairedScanner constructor.

- Fix Logging fatal loop to use direct Serial.printf instead of re-entering Logging::error.

- Remove unused Utils.h include from main.cpp and stale TODO from ScannedCamera.h.

+++ m5stack-core2/lib/utils/Utils.h
@@ -9,7 +9,7 @@ namespace Utils {
 /**
  * Convert uint8 array into hex string.
  */
-std::string hexStr(const uint8_t* data, int len);
+std::string hexStr(const uint8_t* data, size_t len);
 
 /**
  * Convert a given uint32 number into 4 bytes little endian.
